Address code review comment

+++ db/db_impl/db_impl.cc
@@ -4912,7 +4912,9 @@ Status DBImpl::WriteOptionsFile(bool need_mutex_lock,
 
   if (s.ok()) {
     s = RenameTempFileToOptionsFile(file_name);
-  } else {
+  }
+
+  if (!s.ok() && GetEnv()->FileExists(file_name).ok()) {
     if (!GetEnv()->DeleteFile(file_name).ok()) {
       ROCKS_LOG_WARN(immutable_db_options_.info_log,
                      "Unable to delete temp options file %s",
